app/src/main/java/io/pivotal/racquetandroid/fragment/ClubFragment.java
package io.pivotal.racquetandroid.fragment;

import android.os.Bundle;
import android.support.annotation.Nullable;
import android.support.v4.app.Fragment;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;
import android.widget.EditText;
import android.widget.ImageButton;

import butterknife.Bind;
import butterknife.ButterKnife;
import io.pivotal.racquetandroid.R;

public class ClubFragment extends Fragment {
    @Bind(R.id.loser)
    EditText loser;

    @Bind(R.id.winner)
    EditText winner;

    @Bind(R.id.submit)
    ImageButton submit;

    @Nullable
    @Override
    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
        return inflater.inflate(R.layout.fragment_club, null);
    }

    @Override
    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
        super.onViewCreated(view, savedInstanceState);
        ButterKnife.bind(this, view);
    }

    public static ClubFragment newInstance() {
        return new ClubFragment();
    }
}
